Alterações nos atributos em geral

+++ src/java/br/vn/Model/Funcionario.java
@@ -26,7 +26,7 @@ public class Funcionario implements Serializable {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private int id;
+    private long id;
 
     @Column(length = 100)
     private String nome;
@@ -47,13 +47,15 @@ public class Funcionario implements Serializable {
     public Funcionario() {
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
+
+
 
     public String getNome() {
         return nome;
